Finish working in add all items to cart. Important! Read description
Didn't took into consideration products with variants. We'll need to ask the stakeholder if we need to take into consideration variance or not. Also, code for erasing the whole cart is fully functional, but might cause issues in other parts of the software because of the way the catalog pages make an initial GET request when loading. If no issues are found this code is way more efficient that the code to erase each item individually.