Change way to delete items from cart
Change method of deleting items from the cart but still getting an error on page load. The API request for getting the cart when the page load return cart not found if all the items are deleted from the cart. The other option would be to update the quantity of all items in the cart to 0, but this may cause issues related to checkout. Deleting items one-by-one is the best solution at the time. It would be good to revisit this later.